Consolidate citation metrics into vendored evaltools
Replace the inferior vendored has_citation/citation_match metrics with the
better file-extension-aware any_citation/citations_matched metrics (moved from
evals/evaluate.py into evaltools code_metrics.py). Removes the duplicate class
definitions and register_metric calls from evaluate.py so there is a single set.
